# openapi/jewelheart.yaml
openapi: 3.0.3
info:
  title: JewelHeart Admin API
  description: |
    Retreat volunteer scheduling API for JewelHeart mobile apps (iOS/Android).
    Hosted on the same Cloudflare tunnel as KarmaDots: **https://api.karmadots.org**.

    All endpoints except `/jewelheart/health` require a **Firebase ID token** (same as KarmaDots private-server):
    `Authorization: Bearer <firebase_id_token>`.

    **ACL:** The server must verify the token and enforce JewelHeart admin / per-retreat access (implementation detail; not expressed in this contract).

    **Volunteer calendar feeds** (webcal/ICS) and **assignment confirmation** payloads are specified under
    `/jewelheart/calendar-feed/{feedToken}` and `/jewelheart/assignment-confirmations/...` (see tags Calendar, Confirmations).

    **Reminder policy:** native calendar clients apply `VALARM`/alarms from subscribed ICS; do not rely on server push for calendar-specific alerts.
  version: 0.1.0
  license:
    name: Proprietary
servers:
  - url: https://api.karmadots.org
    description: Production (Cloudflare tunnel → private-server)
...
paths:
...
  /jewelheart/volunteers/{volunteerId}/calendar-feed:
    parameters:
      - $ref: "#/components/parameters/VolunteerId"
    post:
      tags: [Calendar]
      summary: Ensure calendar feed exists and return HTTPS + webcal subscribe URLs
      description: >
        Generates `jewelheart_volunteers.calendar_feed_token` when missing (Firebase-authenticated caller with the same ACL as PATCH volunteer).

        Tokens are irrevocable bearer secrets; treat revoked tokens as Gone (410) on the public ICS route via empty lookup.

        `regenerate:true` rotates the opaque token atomically — previous subscribe links stop updating after clients refresh TTL.
      operationId: mintVolunteerCalendarFeed
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VolunteerCalendarFeedMintRequest"
      responses:
        "200":
          description: Current or freshly minted feed metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VolunteerCalendarFeedResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Calendar]
      summary: Revoke calendar feed token for volunteer
      description: Deletes `calendar_feed_token`; existing subscriptions stale out on next publisher refresh cycle.
      operationId: revokeVolunteerCalendarFeed
      responses:
        "204":
          description: Feed revoked (noop if absent)
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /jewelheart/calendar-feed/{feedToken}:
    parameters:
      - $ref: "#/components/parameters/FeedToken"
    head:
      tags: [Calendar]
      summary: Probe feed availability (validators / cache priming)
      description: Mirrors GET caching semantics without transferring the ICS payload.
      operationId: headVolunteerCalendarFeed
      security: []
      responses:
        "200":
          description: Feed exists for token
          headers:
            Cache-Control:
              schema:
                type: string
                description: Typical `private, max-age=0, must-revalidate`; SHOULD mirror GET `ETag`/`Last-Modified`.
            ETag:
              schema:
                type: string
            Last-Modified:
              schema:
                type: string
                format: http-date
        "404":
          $ref: "#/components/responses/NotFound"
        "410":
          description: Former token revoked
    get:
      tags: [Calendar]
      summary: Subscribable ICS document for assignments visible to volunteer
      description: >
        Anonymous bearer-protected endpoint; security depends entirely on `feedToken` entropy.

        **Refresh behavior:** Returning clients poll with `If-None-Match`; respond `304 Not Modified` when assignment set unchanged.

        **Invalidation cues:** Bump `SEQUENCE`/UID stability per assignment UUID (see ICS notes in `jewelheart-calendar-feed-notes.fragment.js`).


        Responses SHOULD include stable `SUMMARY`/slot copy + twin `DISPLAY` alarms (e.g., -PT24H, -PT3H) encoded as `VALARM`
        entries so subscribing calendar apps synthesize alarms without Firebase push notifications.
      operationId: getVolunteerCalendarIcsFeed
      security: []
      responses:
        "200":
          headers:
            Cache-Control:
              schema:
                type: string
            ETag:
              schema:
                type: string
            Last-Modified:
              schema:
                type: string
                format: http-date
          content:
            text/calendar:
              schema:
                type: string
                format: binary
                description: |
                  RFC 5545 document with one `VEVENT` per assignment (JOIN tasks/slots/jobs/retreats). Each event SHOULD include reminders as `VALARM`.
        "304":
          description: Conditional GET satisfied
        "404":
          description: Unknown or mistyped token
        "410":
          description: Token rotated away or volunteer removed
...
components:
  securitySchemes:
    FirebaseBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Firebase Auth ID token
  parameters:
...
    VolunteerId:
      name: volunteerId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    FeedToken:
      name: feedToken
      in: path
      required: true
      description: |
        Opaque high-entropy secret stored on `jewelheart_volunteers.calendar_feed_token`.
        Subscribers may suffix `.ics` or use `webcal://` — server should treat token with or without `.ics` suffix if desired.
      schema:
        type: string
        minLength: 24
...
  responses:
    Unauthorized:
      description: Missing or invalid Firebase token
    NotFound:
      description: Resource not found
  schemas:
...
    VolunteerCalendarFeedMintRequest:
      type: object
      properties:
        regenerate:
          type: boolean
          default: false
    VolunteerCalendarFeedResponse:
      type: object
      required: [subscribeHttpsUrl]
      properties:
        subscribeHttpsUrl:
          type: string
          format: uri
          description: Canonical https://…/jewelheart/calendar-feed/{token} subscriber link.
        webcalSubscribeUrl:
          type: string
          description: Convenience `webcal://` scheme substitution for Apple Calendar onboarding.
          example: webcal://api.karmadots.org/jewelheart/calendar-feed/…
        lastRotatedAt:
          type: string
          format: date-time
...
security:
  - FirebaseBearer: []